+++ src/js/components/Connection.js
@@ -1,34 +1,11 @@
 import React from 'react'
 import { connect } from 'react-redux'
-
-import ChannelStore from '../stores/channel-store'
-
-import ConnectionHeader from './connection-header'
+import ConnectionHeader from './ConnectionHeader'
 import ChannelList from './channel-list'
 import JoinChannel from './join-channel'
+// import Channel from './channel'
 
-import Channel from './channel'
-
-function getCurrentChannel () {
-  return {
-    channel: ChannelStore.getSelectedChannel()
-  }
-}
-
-class MessageCenter extends React.Component {
-  constructor (props) {
-    super(props)
-    this.state = getCurrentChannel()
-  }
-
-  componentWillMount () {
-    ChannelStore.addChangeListener(this._onChange.bind(this))
-  }
-
-  _onChange () {
-    this.setState(getCurrentChannel())
-  }
-
+class Connection extends React.Component {
   render () {
     const {
       connection
@@ -45,7 +22,7 @@ class MessageCenter extends React.Component {
             <JoinChannel />
           </div>
         </div>
-        <Channel channel={this.state.channel} />
+        {/* <Channel channel={this.state.channel} /> */}
       </div>
     )
   }
@@ -58,7 +35,7 @@ export default connect((state, ownProps) => {
   return {
     connection
   }
-})(MessageCenter)
+})(Connection)
 
 function getConnectionById (state, connectionId) {
   return state.connections.list.find(({ id }) => id === connectionId)
