+++ src/js/components/ConnectionHeader.js
