+++ src/js/main.js
@@ -6,7 +6,7 @@ import { createStore, applyMiddleware, compose } from 'redux'
 import loggerMiddleware from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import { rootReducer } from './reducers'
-import ConnectionCreator from './components/connection-creator'
+import ConnectionCreator from './components/ConnectionCreator'
 import MessageCenter from './components/message-center'
 import browserHistory from './modules/browser-history'
 const inProduction = process.env.NODE_ENV === 'production'
