+++ src/js/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux'
 import connections from './connections'
+import conversations from './conversations'
 
 export const rootReducer = combineReducers({
-  connections
+  connections,
+  conversations
 })
